+++ src/apps/projects/apps.py
@@ -1,0 +1,10 @@
+"""Projects app config."""
+
+from django.apps import AppConfig
+
+
+class ProjectsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.projects"
+    label = "projects"
+    verbose_name = "Proyectos"